Fixed Plugin problem
Sooooo basically... the config is littering  your screen with errors when you try to install the config, because my monke ass brain thought, that it would be a nice idea to load the plugins at the bottom of the config, after the plugins get configured. That is dumb, because the plugins dont exist because they arent loaded, and so the packer plugin also doesn't load. Now i moved the plugins into an seperate file and as you can see in line 11 after the cool ascii art, the plugins get loaded first.

This was probably the worst mistake iv'e ever done in a dotfile and it got me like 💀💀💀.